% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WTBr.R
\docType{package}
\name{WTPBr}
\alias{WTPBr}
\alias{WTPBr-package}
\title{WTPBr: A package with...}
\description{
This package is an exercise in package creation using
R studio. The package includes a sample function and
a sample dataset with their respective documentation.
}
\references{
\url{https://github.com/TatianaFerrari/My-First-Repository}
}
\author{
Tatiana Ferrari, School of Geography and Earth Sciences, McMaster University \email{paezha@mcmaster.ca}
}
